Additions: Linux: rcvboxadd: Handle waiting time better in 'reload' and 'status-kernel', bugref:10359.
In case of 'rcvboxadd status-kernel', assuming we use it to check if loaded kernel
modules correspond to GA installation version, we do not need to wait while the check.
It will only waste a few seconds just to figure out that proper modules not loaded.

Waiting is only needed for 'rcvboxadd reload' when modprobe(8) will attempt to (re-)load
modules asynchronously, so we make sure that operation is completed.


git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107908 cfe28804-0f27-0410-a406-dd0f0b0b656f